Move the methods in TestToolkit to TestUtil
