+++ e2e/common/utilities/test-util.ts
@@ -1,10 +1,35 @@
-import { ElementArrayFinder, ElementFinder } from 'protractor';
+import { browser, ElementArrayFinder, ElementFinder } from 'protractor';
 import { ExpectsUtil } from './expects-util';
 import { FormData } from '../components/form.service';
+import { JSConsole } from './js-console';
 
 declare const allure: any;
 
 export class TestUtil extends ExpectsUtil {
+
+	private static console = new JSConsole();
+
+	public static init(tms: string, feature: string, version: string, user: string) {
+		allure.addLabel('tms', tms);
+		allure.addLabel('feature', feature);
+		browser.driver.getCapabilities()
+			.then((caps) => {
+				browser.browserName = caps.get('browserName');
+				allure.addLabel('browser', browser.browserName);
+			});
+		if (version) {
+			allure.addLabel('appVersion', version);
+		}
+		if (user) {
+			allure.addLabel('tester', user);
+		}
+		allure.addLabel('testExecutionDateTime', new Date().toLocaleString());
+		this.console.clear();
+	}
+
+	public static hasErrorsInConsole() {
+		return this.console.hasErrors();
+	}
 
 	public static checkValue(field: ElementFinder, name: string, expectedValue: string) {
 		allure.createStep(name + ' is equals to ' + expectedValue , () => {
